+++ test/data.jl
@@ -245,11 +245,13 @@ end
     df2[!, :A] = ["a", missing, "c"]
     df2[!, :C] = Union{Int, Missing}[1, 2, 4]
 
-    m1 = innerjoin(df1, df2, on = :A)
+    @test_throws ArgumentError innerjoin(df1, df2, on = :A)
+    m1 = innerjoin(df1, df2, on = :A, matchmissing=:equal)
     @test size(m1) == (3,3)
     @test m1[!, :A] ≅ ["a","a", missing]
 
-    m2 = outerjoin(df1, df2, on = :A)
+    @test_throws ArgumentError outerjoin(df1, df2, on = :A)
+    m2 = outerjoin(df1, df2, on = :A, matchmissing=:equal)
     @test size(m2) == (5,3)
     @test m2[!, :A] ≅ ["a", "b", "a", missing, "c"]
 end
@@ -264,12 +266,12 @@ end
                     v2 = 1:6)
     df2[1,:a] = missing
 
-    m1 = innerjoin(df1, df2, on = [:a,:b])
+    m1 = innerjoin(df1, df2, on = [:a,:b], matchmissing=:equal)
     @test m1 == DataFrame(a=[:x,:x,:x,:x,:x,:y,:x,:x],
                           b=[:A,:A,:A,:A,:B,:B,:A,:A],
                           v1=[1,1,2,2,3,4,5,5],
                           v2=[3,6,3,6,4,2,3,6])
-    m2 = outerjoin(df1, df2, on = [:a,:b])
+    m2 = outerjoin(df1, df2, on = [:a,:b], matchmissing=:equal)
     @test m2 ≅ DataFrame(a=[:x,:x,:x,:x,:x,:y,:x,:x,:x,missing,:x],
                          b=[:A,:A,:A,:A,:B,:B,:A,:A,:D,:A,:C],
                          v1=[1,1,2,2,3,4,5,5,6,missing,missing],
@@ -481,6 +483,7 @@ end
     @test filter(Symbol[] => flipflop0, df) == df[[1,3], :]
     @test filter(r"z" => flipflop0, df) == df[[1,3], :]
     @test filter(Not(All()) => flipflop0, df) == df[[1,3], :]
+    @test filter(Cols() => flipflop0, df) == df[[1,3], :]
     @test filter(AsTable(r"z") => flipflop1, df) == df[[1,3], :]
     @test filter(AsTable([]) => flipflop1, df) == df[[1,3], :]
     @test filter!([] => flipflop0, copy(df)) == df[[1,3], :]
@@ -489,6 +492,7 @@ end
     @test filter!(Symbol[] => flipflop0, copy(df)) == df[[1,3], :]
     @test filter!(r"z" => flipflop0, copy(df)) == df[[1,3], :]
     @test filter!(Not(All()) => flipflop0, copy(df)) == df[[1,3], :]
+    @test filter!(Cols() => flipflop0, copy(df)) == df[[1,3], :]
     @test filter!(AsTable(r"z") => flipflop1, copy(df)) == df[[1,3], :]
     @test filter!(AsTable([]) => flipflop1, copy(df)) == df[[1,3], :]
 
@@ -504,6 +508,7 @@ end
         @test names(v, Between(:x1, :x3)) == ["x1", "x2", "x3"]
         @test names(v, Not(:a)) == names(v, r"x") == ["x1", "x2", "x3", "x4"]
         @test names(v, :x1) == names(v, 2) == ["x1"]
+        @test names(v, Cols()) == names(v, Cols()) == []
     end
 
     for v in [view(df, :, [4,3,2,1]), groupby(view(df, :, [4,3,2,1]), 1), view(df, 1, [4,3,2,1])]
@@ -511,6 +516,7 @@ end
         @test names(v, Between(:x2, :x1)) == ["x2", "x1"]
         @test names(v, Not(:a)) == names(v, r"x") == ["x3", "x2", "x1"]
         @test names(v, :x1) == names(v, 3) == ["x1"]
+        @test names(v, Cols()) == names(v, Cols()) == []
     end
 end